Fix grpc issues and nsqd tests
This commit makes the invoke() command in the Checker actually safe. It
fixes incorrect assumptions that were made about the way the reflection
stuff works. Also, we now correctly clean up NSQd before and after every
test that uses it. This necessitated a change in build-go to only run
tests serially. We will at some point have to fix this.