Gozer now stores state in an external volume
In order to simplify the systemd dependency tree and allow the connector
to gracefully fail, we must remove the dependency on using
--volumes-from against the connector container. By making gozer store
state in an external volume, we can easily do this. The register service
will have to be made resilient to its ip state file going missing, but
for now the register service can just crash if the file goes missing.